Move toaster to chakra folder, set default sceanrio Id when scenario from url is not valid
