+++ src/components/chakra/provider.tsx
@@ -5,7 +5,7 @@ import {
   type ColorModeProviderProps,
 } from "./color-mode";
 import { system } from "./theme";
-import { Toaster } from "../ui/toaster";
+import { Toaster } from "./toaster";
 
 export function Provider(props: ColorModeProviderProps) {
   return (
